src/ecs/insights_hub/backend_python/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
boto3>=1.34.0
python-dotenv>=1.0.0
typing-extensions>=4.9.0
httpx>=0.26.0
pytest>=7.4.0
python-multipart>=0.0.6
opensearch-py>=2.8.0
pydantic>=2.0.0
strands-agents
strands-agents-tools